+++ library/JQLibrary/src/JQGuetzli/guetzli/jpeg_data_reader.cc
@@ -465,7 +465,7 @@ struct BitReaderState {
     FillBitWindow();
     uint64_t val = (val_ >> (bits_left_ - nbits)) & ((1ULL << nbits) - 1);
     bits_left_ -= nbits;
-    return val;
+    return static_cast<int>(val);
   }
 
   // Sets *pos to the next stream position where parsing should continue.
