+++ library/JQLibrary/src/JQQRCodeReader/zxing/zxing/oned/CodaBarReader.cpp
@@ -77,7 +77,7 @@ CodaBarReader::CodaBarReader()
 Ref<Result> CodaBarReader::decodeRow(int rowNumber, Ref<BitArray> row) {
 
   { // Arrays.fill(counters, 0);
-    int size = counters.size();
+    int size = static_cast< int >( counters.size() );
     counters.resize(0);
     counters.resize(size); }
 
@@ -168,7 +168,7 @@ void CodaBarReader::validatePattern(int start)  {
   // First, sum up the total size of our four categories of stripe sizes;
   vector<int> sizes (4, 0);
   vector<int> counts (4, 0);
-  int end = decodeRowResult.length() - 1;
+  int end = static_cast< int >( decodeRowResult.length() ) - 1;
 
   // We break out of this loop in the middle, in order to handle
   // inter-character spaces properly.
