+++ library/JQLibrary/src/JQQRCodeReader/zxing/zxing/oned/Code128Reader.cpp
@@ -185,7 +185,7 @@ vector<int> Code128Reader::findStartPattern(Ref<BitArray> row){
   vector<int> counters (6, 0);
   int patternStart = rowOffset;
   bool isWhite = false;
-  int patternLength =  counters.size();
+  int patternLength = static_cast< int >( counters.size() );
 
   for (int i = rowOffset; i < width; i++) {
     if (row->get(i) ^ isWhite) {
@@ -311,7 +311,7 @@ Ref<Result> Code128Reader::decodeRow(int rowNumber, Ref<BitArray> row) {
 
     // Advance to where the next code will to start
     lastStart = nextStart;
-    for (int i = 0, e = counters.size(); i < e; i++) {
+    for (int i = 0, e = static_cast< int >( counters.size() ); i < e; i++) {
       nextStart += counters[i];
     }
 
@@ -499,7 +499,7 @@ Ref<Result> Code128Reader::decodeRow(int rowNumber, Ref<BitArray> row) {
   }
 
   // Need to pull out the check digits from string
-  int resultLength = result.length();
+  int resultLength = static_cast< int >( result.length() );
   if (resultLength == 0) {
     // false positive
     throw NotFoundException();
@@ -518,7 +518,7 @@ Ref<Result> Code128Reader::decodeRow(int rowNumber, Ref<BitArray> row) {
   float left = (float) (startPatternInfo[1] + startPatternInfo[0]) / 2.0f;
   float right = lastStart + lastPatternSize / 2.0f;
 
-  int rawCodesSize = rawCodes.size();
+  int rawCodesSize = static_cast< int >( rawCodes.size() );
   ArrayRef<char> rawBytes (rawCodesSize);
   for (int i = 0; i < rawCodesSize; i++) {
     rawBytes[i] = rawCodes[i];
