+++ library/JQLibrary/src/JQQRCodeReader/zxing/zxing/oned/EAN13Reader.cpp
@@ -45,7 +45,7 @@ int EAN13Reader::decodeMiddle(Ref<BitArray> row,
   for (int x = 0; x < 6 && rowOffset < end; x++) {
     int bestMatch = decodeDigit(row, counters, rowOffset, L_AND_G_PATTERNS);
     resultString.append(1, (char) ('0' + bestMatch % 10));
-    for (int i = 0, end = counters.size(); i <end; i++) {
+    for (int i = 0, end = static_cast< int >( counters.size() ); i <end; i++) {
       rowOffset += counters[i];
     }
     if (bestMatch >= 10) {
@@ -62,7 +62,7 @@ int EAN13Reader::decodeMiddle(Ref<BitArray> row,
     int bestMatch =
       decodeDigit(row, counters, rowOffset, L_PATTERNS);
     resultString.append(1, (char) ('0' + bestMatch));
-    for (int i = 0, end = counters.size(); i < end; i++) {
+    for (int i = 0, end = static_cast< int >( counters.size() ); i < end; i++) {
       rowOffset += counters[i];
     }
   }
